added task to generate stubs and removed handcrafted stubs
