+++ sapphire/examples/hanksTodoRuby/build.gradle
@@ -7,7 +7,16 @@ apply plugin: 'java'
 */
 // Customize app stub generation for this example - disable it for now - see above.
 genStubs {
-    onlyIf() { false } // Disable
+    main = "sapphire.compiler.GraalStubGenerator"
+    classpath =  files("$projectDir/../../sapphire-core/build/libs/sapphire-core-1.0.0.jar")
+
+    def sapphireObjectPath = "$projectDir/src/main/js/sapphire/appexamples/hankstodo/todo_list_manager.js"
+    def outPath = "$projectDir/src/main/java"
+    def packageName = "sapphire.appexamples.hankstodo.stubs"
+    def sapphireClasses = "TodoListManager"
+    args sapphireObjectPath, outPath, packageName, sapphireClasses
+    outputs.dir "$projectDir/src/main/java/sapphire/appexamples/hankstodo/stubs" // Declare outputs, so gradle will run if they have been changed
+    inputs.dir sapphireObjectPath   // Declare inputs, so gradle will run if they have been changed
 }
 // Customize stub compilation for this example
 compileStubs {
@@ -34,8 +43,8 @@ task runjsapp(type: Exec) {
     environment  JS_HOME: "${project.projectDir}/src/main/js/sapphire/appexamples/hankstodo"
     environment  OMS_IP: "${project.property('omsIp')}"
     environment  OMS_PORT: "${project.property('omsPort')}"
-    environment  KERNEL_SERVER_IP: "${project.property('embeddedKernelServerIp')}"
-    environment  KERNEL_SERVER_PORT: "${project.property('kernelServerPort')}"
+    environment  HOST_IP: "${project.property('embeddedKernelServerIp')}"
+    environment  HOST_PORT: "${project.property('kernelServerPort')}"
     def classpath = sourceSets.main.runtimeClasspath.asPath
     executable "$System.env.JAVA_HOME/bin/js"
     args "--jvm", "--polyglot", "--jvm.classpath=${classpath}",
@@ -46,8 +55,9 @@ task runjsapp(type: Exec) {
 runapp {
     systemProperty "RUBY_HOME",  "${project.projectDir}/src/main/ruby/sapphire/appexamples/hankstodo/"
     systemProperty "JS_HOME",  "${project.projectDir}/src/main/js/sapphire/appexamples/hankstodo/"
-    main = 'sapphire.appexamples.hankstodo.stubs.Client_Stub'
+    main = 'sapphire.appexamples.hankstodo.Client'
     args = [project.property('embeddedKernelServerIp'), project.property('kernelServerPort'), project.property('omsIp'), project.property('omsPort')]
 }
 
+compileJava.dependsOn genStubs
 runapp.dependsOn compileStubs
