+++ sapphire/examples/hanksTodoRuby/src/main/js/sapphire/appexamples/hankstodo/hanks_todo_main.js
@@ -1,17 +1,12 @@
 function main() {
     system = Java.type("java.lang.System")
-    var jsHome = system.getenv("JS_HOME")
-    var kernelServerIP = system.getenv("KERNEL_SERVER_IP")
-    var kernelServerPort = system.getenv("KERNEL_SERVER_PORT")
+    var hostIP = system.getenv("HOST_IP")
+    var hostPort = system.getenv("HOST_PORT")
     var omsIP = system.getenv("OMS_IP")
     var omsPort = system.getenv("OMS_PORT")
 
-    system.setProperty("JS_HOME", jsHome)
-    var fileName = jsHome + "/todo_list_manager.js"
-
-    var ClientStub = Java.type("sapphire.appexamples.hankstodo.stubs.Client_Stub")
-    var clientStub = new ClientStub()
-    var tlm = clientStub.GetTodoListManager(kernelServerIP, kernelServerPort, omsIP, omsPort, "js", fileName);
+    var ClientStub = Java.type("sapphire.appexamples.hankstodo.stubs.TodoListManager_Stub")
+    var tlm = ClientStub.getStub("HanksTodoJS.yaml", omsIP, omsPort, hostIP, hostPort);
 
     var t1 = tlm.newTodoList("Hanks")
     console.log(tlm.addTodo("Hanks", "First todo"));
