+++ sapphire/examples/hanksTodoRuby/src/main/resources/HanksTodoRuby.yaml
@@ -1,0 +1,10 @@
+!!sapphire.app.SapphireObjectSpec
+constructorName: TodoListManager
+dmList:
+- configs:
+  - !!sapphire.policy.dht.DHTPolicy$Config {numOfShards: 2}
+  name: sapphire.policy.dht.DHTPolicy
+javaClassName: sapphire.appexamples.hankstodo.stubs.TodoListManager_Stub
+lang: ruby
+name: TodoListManagery
+sourceFileLocation: ./src/main/ruby/sapphire/appexamples/hankstodo/todo_list_manager.rb
